rails40 - rspec running but failing
Controller and routing specs, predictable. Probably not too
difficult to fix, but not in this commit.

The secret for devise in this commit for temporary convenience.
It will be changed and pulled out of a later commit. This code
won't ever be deployed in production, so it doesn't really matter
that much, but it's good to build good habits.