rails41 - using rails:update rake task on this branch
For whatever reason, I've not used or forgotten about the
rails update task recently. It's time to start rolling an
invocation of that into my regular upgrade procedure. Not
every file has to be changed, but it's super useful for
keeping up with new files added to rails applications, and
for watching how configurations change.