rails41 - more rails:update additions
Many of the updates simply do a dumb overwrite of existing
and important configuration information, so it's not as
simple as just running the command and letting it rip.
This commit chews away at a bit more of it. There is probably
one more commit for dealing with environment file updates,
then it should be good to merge back into master.