rails41 - fixed some pending specs
For whatever reason, some specs which were not working
and moved to pending are now working again. Very weird,
but I'll take it.

There is almost surely more to do with the application
update task, but it will take a bit of focus on that
specifically, and should be in an upcoming commit.