rails42: configuration files update
Most of the changes here revolve around squaring
the various files in `config/` with the rails 4.2
version of `rake rails:update`. Nost of this is
unused with the currect application, but it feels
important to know for going forward.

It's the middle of May and the fog hasn't started
in yet. Pretty crazy.