more rspec, gem and rubocop update
Another version bump for rubocop, this one is going
to require changing some namespaces in the .rubocop.yml
configuration. Not for this commit.

The features are still broken, need to spend some time
unscrewing cucumber.

There is also an asset problem with blueprint css.
This would be a good time to change over to bootstrap,
should be done on another branch though.