cucumber: much devise updating here
While the devise gem had been updated, the devise
implementation was lagging. That's all be updated
in this commit, which is a pretty big one. Ended up
redoing all the devise migrations as well. Not
something advisable for a production application,
but no big deal here, as the point is to get to the
cucumber bdd work asap.

The specs still run, cucumber still fails in exactly
the same place.

However, it's now possible to actually log into the
running application, which is progress.